Big Host Display Update
Added Arrays to hold the table data and a 2D array to hold the arrays.
The fields now populate on start up from these arrays. The activities
dialog now only has two buttons that toggle whether a table is occupied
or not and make reservation. Atif change the for loop index name to
tableNo which makes more sense. A couple updates on the field error
checking for text input. Also changed the edit pane buttons so there’s
just a done button instead of a close dialog and save button. They both
did the same thing anyway.